fix: w3c adaptive
