+++ src/components/others/BurgerBtn/BurgerBtn.tsx
@@ -8,15 +8,19 @@ interface IBurgerBtnProps {
 }
 
 const BurgerBtn: React.FC<IBurgerBtnProps> = ({ onClick, isOpen }) => (
-  <button
+  <div
     className={`burger-menu btn-hover btn-reset ${isOpen ? 'burger-menu_open' : ''}`}
+    role='button'
     onClick={onClick}
-    type='button'
+    tabIndex={0}
+    onKeyDown={(e) => {
+      if (e.code === 'Enter') onClick();
+    }}
   >
     <div className='burger-menu__line burger-menu__line_type_top' />
     <div className='burger-menu__line burger-menu__line_type_middle' />
     <div className='burger-menu__line burger-menu__line_type_bottom' />
-  </button>
+  </div>
 );
 
 export default BurgerBtn;
