+++ src/index.tsx
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
 // eslint-disable-next-line import/extensions
-import App from './components/App/App';
+import App from './components/common/App/App';
 
 // import App from './components/App/App';
